backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.44.0
google-cloud-secret-manager==2.16.3
google-cloud-firestore==2.14.0
google-cloud-pubsub==2.18.4
google-cloud-run==0.10.3
python-dotenv==1.0.0
